Improve performance by doing less work
This commit does a couple of things to improve performance of the
library.

First, it doesn't calculate all the default options in the
utility getClassName function. Instead, it only calculates the ones
that weren't set in the call. And that logic was moved into the mixin
(which is also used by the decorator) so that direct calls to utility
or the HOC approach don't look for all those defaults on `this`.

Second, it configures the base component name for the HOC when the
HOC is defined instead of on each invocation.